[Refactor]: 방 조회 관련 폴더 구조 변경 및 컴포넌트로 리팩터링 - 전체 채팅방, 내가 속한 채팅방만 조회 가능하도록 변경
