+++ next/src/app/(home)/(communication)/channel/layout.tsx
@@ -1,7 +1,8 @@
 import childrenProps from '@/interfaces/childrenProps.interface';
 import { Metadata } from 'next';
-import ChatRoomList from '@/app/(home)/(communication)/channel/_list/chatRoomList';
-import RoomBuilder from '@/app/(home)/(communication)/channel/roomBuilder';
+import ChatRoomList from '@/app/(home)/(communication)/channel/_room/chatRoomList';
+import RoomBuilder from '@/app/(home)/(communication)/channel/_room/roomBuilder';
+import MyChatRoomList from '@/app/(home)/(communication)/channel/_room/myChatRoomList';
 
 export const metadata: Metadata = {
   title: 'channel',
@@ -14,6 +15,7 @@ export default function ChannelLayout({ children }: childrenProps) {
       <h1>Channel</h1>
       <div style={{ display: 'flex', gap: '200px' }}>
         <ChatRoomList />
+        <MyChatRoomList />
         <RoomBuilder title={'방 생성'} method={'POST'} url={'chat'} />
         {children}
       </div>
